# docs/openapi.yml
# THIS FILE IS GENERATED BY OpenApiSchemaExportTest — DO NOT EDIT MANUALLY.
# Run `./mvnw -pl service -am test -Dtest=OpenApiSchemaExportTest -Dsurefire.failIfNoSpecifiedTests=false` to regenerate.
---
openapi: 3.1.0
components:
  schemas:
    ApplicationSpec:
      type: object
      deprecated: true
      required:
      - repoId
      - name
      properties:
        repoId:
          type: string
          pattern: \S
        name:
          type: string
          pattern: \S
        healthPath:
          type: string
    CreateEnvironmentRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          pattern: \S
        branch:
          type: string
        network:
          type: string
        applications:
          type: array
          deprecated: true
          items:
            $ref: "#/components/schemas/ApplicationSpec"
    EnvironmentResponse:
      type: object
      properties:
        environment:
          $ref: "#/components/schemas/PdEnvironmentDto"
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    ListApplicationsResponse:
      type: object
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/PdApplicationDto"
    ListDeploymentsResponse:
      type: object
      properties:
        deployments:
          type: array
          items:
            $ref: "#/components/schemas/PdDeploymentDto"
    ListEnvironmentsResponse:
      type: object
      properties:
        environments:
          type: array
          items:
            $ref: "#/components/schemas/PdEnvironmentDto"
    ListLinksResponse:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/PdLinkedServiceDto"
    ListPinsResponse:
      type: object
      properties:
        pins:
          type: array
          items:
            $ref: "#/components/schemas/PdPinDto"
    ListServicesResponse:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/PdServiceDto"
    PdApplicationDto:
      type: object
      properties:
        id:
          type: string
        repoId:
          type: string
        name:
          type: string
        environmentId:
          type: string
        environmentName:
          type: string
        target:
          $ref: "#/components/schemas/PdDeploymentTarget"
        availableOnEnv:
          type: boolean
        branch:
          type: string
        healthPath:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
    PdDeploymentDto:
      type: object
      properties:
        id:
          type: string
        applicationId:
          type: string
        applicationName:
          type: string
        commitSha:
          type: string
        runId:
          type: string
        status:
          $ref: "#/components/schemas/PdDeploymentStatus"
        containerName:
          type: string
        detail:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        finishedAt:
          $ref: "#/components/schemas/Instant"
    PdDeploymentStatus:
      type: string
      enum:
      - QUEUED
      - STARTING
      - ACTIVE
      - IMAGE_MISSING
      - FAILED
      - DECOMMISSIONED
    PdDeploymentTarget:
      type: string
      enum:
      - ENVIRONMENT
      - PLATFORM
    PdEnvironmentDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        branch:
          type: string
        network:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        applications:
          type: array
          items:
            $ref: "#/components/schemas/PdApplicationDto"
    PdLinkedServiceDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        target:
          $ref: "#/components/schemas/PdDeploymentTarget"
        availableOnEnv:
          type: boolean
        healthPath:
          type: string
    PdPinDto:
      type: object
      properties:
        applicationName:
          type: string
        shas:
          type: array
          items:
            type: string
    PdServiceDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        target:
          $ref: "#/components/schemas/PdDeploymentTarget"
        branch:
          type: string
        availableOnEnv:
          type: boolean
        healthPath:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        environmentIds:
          type: array
          items:
            type: string
    UpdateEnvironmentRequest:
      type: object
      properties:
        name:
          type: string
        branch:
          type: string
    UpsertServiceRequest:
      type: object
      properties:
        deploymentTarget:
          $ref: "#/components/schemas/PdDeploymentTarget"
        branch:
          type: string
        availableOnEnv:
          type: boolean
        healthPath:
          type: string
        environmentIds:
          type: array
          items:
            type: string
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://qits-platform-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
  /platform-deployments/api/applications:
    get:
      summary: Every application deployed here — environment applications and platform
        services
      responses:
        "200":
          description: "The applications, one entry per tier"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListApplicationsResponse"
      tags:
      - Pd Application Controller
  /platform-deployments/api/deployments:
    get:
      summary: "One plane's recorded deployments, newest-first — an environment id,\
        \ or 'platform'"
      parameters:
      - name: environmentId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: The deployments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListDeploymentsResponse"
        "400":
          description: environmentId was not given
        "404":
          description: No such environment
      tags:
      - Pd Deployment Controller
  /platform-deployments/api/environments:
    get:
      summary: List environments
      responses:
        "200":
          description: The environments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvironmentsResponse"
      tags:
      - Pd Environment Controller
    post:
      summary: "Create an environment: a name, a branch to listen to, a bundle network"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnvironmentRequest"
        required: true
      responses:
        "201":
          description: Created; green builds on the branch now deploy
        "400":
          description: "A name, branch or network failed validation"
        "409":
          description: An environment of that name already exists
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Environment Controller
  /platform-deployments/api/environments/{environmentId}:
    patch:
      summary: Rename an environment or point it at another branch
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnvironmentRequest"
        required: true
      responses:
        "200":
          description: The updated environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
        "400":
          description: A name or branch failed validation
        "404":
          description: No such environment
        "409":
          description: Another environment already has that name
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Environment Controller
    get:
      summary: One environment with the applications it tracks
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
        "404":
          description: No such environment
      tags:
      - Pd Environment Controller
    delete:
      summary: "Tear an environment down (rows, containers, networks)"
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Torn down
        "404":
          description: No such environment
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Environment Controller
  /platform-deployments/api/environments/{environmentId}/links:
    get:
      summary: "Every service present in this environment: its links, plus every platform\
        \ service"
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The services present in this environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListLinksResponse"
        "404":
          description: No such environment
      tags:
      - Pd Environment Controller
  /platform-deployments/api/pins:
    get:
      summary: "The image shas deployments pin: what serves, and what a rollback restores"
      responses:
        "200":
          description: One entry per application name serving somewhere; an application
            with no ACTIVE deployment pins nothing and is absent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListPinsResponse"
      tags:
      - Pd Pin Controller
  /platform-deployments/api/services:
    get:
      summary: "Every service, with the environments each is linked into"
      responses:
        "200":
          description: The services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListServicesResponse"
      tags:
      - Pd Service Controller
  /platform-deployments/api/services/{name}:
    put:
      summary: "Register or update one service, replacing its environment links"
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertServiceRequest"
        required: true
      responses:
        "200":
          description: The updated service
          content:
            application/json:
              schema: {}
        "201":
          description: The newly registered service
        "400":
          description: "Validation failed, or a platform service was given environment\
            \ links"
        "404":
          description: An environmentId names no environment
        "409":
          description: A platform service cannot become an environment service
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Service Controller
    delete:
      summary: Remove a service and its links
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Removed
        "404":
          description: No such service
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Service Controller
info:
  title: qits-platform-deployments API
  version: 2026.0.0